Add linter waiver file
This silences the linter for all known warnings. That eases the
migration to a linter flow, because new linter warnings are identified
quickly.

Anyhow, I suggest to curate the waiver file by either solving the
underlying problem or replacing the comment with a meaningful
description.

Also, we can now be pedantic, linter will fail on any warning.

Signed-off-by: Stefan Wallentowitz <stefan.wallentowitz@hm.edu>